% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ANPPmax.R
\name{calc_ANPPmax}
\alias{calc_ANPPmax}
\title{Maximum Above-ground Net Primary Production}
\usage{
calc_ANPPmax(RAIN, MAT, SAND)
}
\arguments{
\item{RAIN}{Long-term Mean Annual Precipitation (mm/year).}

\item{MAT}{Long-term Mean Annual Precipitation (degrees Celsius).}

\item{SAND}{Sand \% in top 30 cm soil}
}
\description{
This function calculates the max ANPP  given mean annual precipitation and temperature.
}
