% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_BNPPest.R
\name{calc_BNPPest}
\alias{calc_BNPPest}
\title{Estimated Below-ground Net Primary Production}
\usage{
calc_BNPPest(
  RAIN,
  MAT,
  ANPPt_est,
  Sk,
  S0 = 0.1 * Sk,
  APCcorrection = FALSE,
  DEPTH = 30
)
}
\arguments{
\item{RAIN}{Long-term Mean Annual Precipitation (mm/year).}

\item{MAT}{Long-term Mean Annual Precipitation (degrees Celsius).}

\item{ANPPt_est}{(equivalent to Pg) Total seasonal grass production for the given grazing management system. Output of calc_ANPPest() or calc_Pg().}

\item{Sk}{The steady state of biomass in the absence of grazing for a given location. This should ideally be measured directly using grazing exclosures.}

\item{S0}{Biomass condition prior to the growing season (at the end of the dry season) that is mostly comprised of carbon stores in rhizomes. Default is 0.1*SK.}

\item{APCcorrection}{Default = FALSE. A correction factor for the influence of annual versus perennial plant growth strategies on belowground production. If correction is applied, then APC = 0.291. This is where forage is dominated by annuals (or shrubs often associated with annuals in drier grasslands). Otherwise, APC = 1.}

\item{DEPTH}{Default = 30. Depth of soil sampling / estimation (cm). The original SNAP model was developed based on measurements to a depth of 40 cm.}
}
\description{
This function calculates the estimated BNPP given mean annual precipitation, temperature, grazing management and the dominance of annual or perennial grasses.
}
