% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Cg.R
\name{calc_Cg}
\alias{calc_Cg}
\title{Daily biomass consumption during the growing season}
\usage{
calc_Cg(W)
}
\arguments{
\item{W}{Average animal body size (kg live weight)}
}
\description{
Calculation of the parameter Cg, the daily biomass consumption rate during the growing season, from Ritchie (2020)
}
