% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_deltaSOC.R
\name{calc_deltaSOC}
\alias{calc_deltaSOC}
\title{Delta SOC}
\usage{
calc_deltaSOC(
  PDSOCt,
  DDSOCt,
  SAND,
  RAIN,
  Gdays,
  SOC,
  lowSOC = FALSE,
  orig = FALSE
)
}
\arguments{
\item{PDSOCt}{Output of calc_PDSOCt()}

\item{DDSOCt}{Output of calc_DDSOCt()}

\item{SAND}{Sand \% in top 30 cm soil}

\item{RAIN}{MAP for year t (mm/year)}

\item{Gdays}{Total number of days in the growing season. Default = 153 (October to March-ish).}

\item{SOC}{Starting soil carbon stocks (g/m2).}

\item{lowSOC}{Default = FALSE. Different regression equation for respiration rate is applied for low and high SOC to avoid a negative respiration rate (which isn't physically possible). Threshold for what qualifies as "low SOC" is 4,600 gC/m^2 (i.e. 46 t/ha). Low SOC regression equation is applicable for higher SOC, but just with slightly lower R-squared.}

\item{orig}{Default = FALSE. Use the original DMRESP equations from Ritchie 2020 or the updated ones from Ruan deWet}
}
\description{
To calculate the change in SOC for year t (deltaSOCt), we need the combination of PDSOCt and DDSOCt, but also the maximum rate of microbial respiration for year t (MRESPt). A key input to MRESPt is WETDAYS, which is calculated as part of this function.
}
