% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dmax.R
\name{calc_dmax}
\alias{calc_dmax}
\title{Maximum sustainable stocking density}
\usage{
calc_dmax(Sf, Sk, Cg, Gdays = 153)
}
\arguments{
\item{Sf}{Biomass at the end of the growing season. Output of calc_Sf.}

\item{Sk}{The steady state of biomass in the absence of grazing for a given location. This should ideally be measured directly using grazing exclosures.}

\item{Cg}{Daily consumption rate (g/animal/day)}

\item{Gdays}{Total number of days in the growing season. Default = 153 (October to March-ish).}
}
\description{
Calculation of the theoretical maximum stocking density, from Ritchie (2020). This is based on the assumptions about the consumption rate in the non-growing season relative to the available biomass.
}
