% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Fdays.R
\name{calc_Fdays}
\alias{calc_Fdays}
\title{Number of growing season days after the grazing event.}
\usage{
calc_Fdays(Gdays, Edays, Ddays)
}
\arguments{
\item{Gdays}{Total number of days in the growing season.}

\item{Edays}{Number of growing season days before the grazing event.}

\item{Ddays}{Number of days of grazing event.}
}
\description{
Number of growing season days after the grazing event.
}
