% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Gdays.R
\name{calc_Gdays}
\alias{calc_Gdays}
\title{Length of growing season}
\usage{
calc_Gdays(RAIN, MAT)
}
\arguments{
\item{RAIN}{Mean annual precipitation (mm/yr)}

\item{MAT}{Mean annual temperature (degrees Celsius)}
}
\description{
Calculation of the length of the growing season in days from Ritchie (2020)
}
