% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gi.R
\name{calc_gi}
\alias{calc_gi}
\title{Grazing intensity}
\usage{
calc_gi(Sf, Lo, Sk, Lg, ANPPest, method)
}
\arguments{
\item{Sf}{Biomass at the end of the growing season. Output of calc_Sf.}

\item{Lo}{Biomass consumed during the offseason. Output of calc_Lo.}

\item{Sk}{The steady state of biomass in the absence of grazing for a given location. Either measured directly or calculated as output of calc_ANPPmax.}

\item{Lg}{Biomass consumed during the growing season. Output of calc_LG.}

\item{ANPPest}{Aboveground net primary productivity during the growing season. Output of calc_ANPPest.}

\item{method}{Method used to calculate grazing intensity. Options are 'Sk' or 'ANPPest'. 'Sk' calculates grazing intensity as the proportional difference in plant biomass between grazed and ungrazed treatments at the end of the offseason. 'ANPPest' calculates grazing intensity as the amount of biomass consumed during both the growing and offseason relative to the amount of biomass produced.}
}
\description{
Calculation of the parameter gi, the grazing intensity of the specified grazing management practices
}
