% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_LG.R
\name{calc_Lg}
\alias{calc_Lg}
\title{Biomass loss during grazing episode}
\usage{
calc_Lg(Ddays, d, n, W = NA, Cg = NA)
}
\arguments{
\item{Ddays}{Number of days of grazing episode}

\item{d}{Stocking density (head/ha)}

\item{n}{Number of "pastures" per total area, A.}

\item{W}{Average animal body size (kg live weight)}

\item{Cg}{Daily consumption rate (g/animal/day)}
}
\description{
Calculation of the parameter Lg, the biomass loss during grazing episode, from Ritchie (2020)
}
