% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Lo.R
\name{calc_Lo}
\alias{calc_Lo}
\title{Loss of biomass during the off season}
\usage{
calc_Lo(Cg, Co = NA, Gdays, DdaysO = NA, d_off)
}
\arguments{
\item{Cg}{Daily biomass consumption rate during the growing season (g/animal/day).}

\item{Co}{Daily biomass consumption rate during the offseason (g/animal/day).}

\item{Gdays}{Total number of days in the growing season.}

\item{DdaysO}{Number of days in offseason grazing episode. Default = 365 - Gdays.}

\item{d_off}{Stocking density during the offseason (head/ha)}
}
\description{
Estimating the amount of biomass lost during the dormant season (dry, non-growing season).
}
