% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_PDSOCt.R
\name{calc_PDSOCt}
\alias{calc_PDSOCt}
\title{Plant-Derived SOC in year t}
\usage{
calc_PDSOCt(BNPPt_est, Sf, Lo, LIGCELL, FIRE)
}
\arguments{
\item{BNPPt_est}{Output of calc_BNPPest()}

\item{Sf}{Biomass at the end of the growing season. Output of calc_Sf() (g/m2)}

\item{Lo}{Loss of biomass during the non-growing season. Output of calc_Lo() (g/m2)}

\item{LIGCELL}{Lignin and cellulose fraction of plant biomass (0 - 1)}

\item{FIRE}{Proportion of biomass consumed by fire during the non-growing season (0-1)}
}
\description{
Plant-derived SOC inputs for year t (PDSOCt) (g/m2)
}
