% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_SE.R
\name{calc_SE}
\alias{calc_SE}
\title{Biomass at the start of grazing episode}
\usage{
calc_SE(Sk, Edays, S0 = 0.1 * Sk, r = 0.05)
}
\arguments{
\item{Sk}{The steady state of biomass in the absence of grazing for a given location. This should ideally be measured directly using grazing exclosures.}

\item{Edays}{Number of days within the growing season prior to grazing episode}

\item{S0}{The initial biomass condition prior to the growing season (at the end of the dry season) that is mostly comprised of carbon stores in rhizomes. Default is 0.1*SK.}

\item{r}{Relative growth rate of grass biomass, which is variable over the growing season given it's density dependence. Can be calculated as tht intercept of a relationship between the measured relative growth rate and biomass at a given time. Default for tropical grasslands is 0.05 and for temperate grasslands is 0.035.}
}
\description{
Calculation of the parameter Se, the biomass at the start of grazing episode, from Ritchie (2020)
}
