% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_SOCeq.R
\name{calc_SOCeq}
\alias{calc_SOCeq}
\title{Equilibrium SOC}
\usage{
calc_SOCeq(PDSOCt, DDSOCt, DEPTH, SAND, RAIN, Gdays, orig = FALSE)
}
\arguments{
\item{PDSOCt}{Output of calc_PDSOCt() (g/m2)}

\item{DDSOCt}{Output of calc_DDSOCt() (g/m2)}

\item{DEPTH}{Soil depth at which equilibrium SOC stock is reported (cm)}

\item{SAND}{Sand fraction in the top 30 cm of soil (0 - 1)}

\item{RAIN}{Mean annual precipitation for year t (mm/year)}

\item{Gdays}{Total number of days in the growing season.}

\item{orig}{Default = FALSE. Use the original DMRESP equations from Ritchie 2020 or the updated ones from Ruan deWet}
}
\description{
Calculate the SOC stock in tSOC/ha at equilibrium under the specified conditions. This is the point at which SOC inputs from plant and dung equal SOC outputs from soil respiration (deltaSOC = 0).
}
