% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNAPGRAZE_delta_ann.R
\name{SNAPGRAZE_delta_ann}
\alias{SNAPGRAZE_delta_ann}
\title{SNAPGRAZE delta Wrapper Function}
\usage{
SNAPGRAZE_delta_ann(
  SAND,
  RAIN,
  MAT,
  FIRE,
  LIGCELL,
  years,
  SOC,
  Sk = NA,
  S0 = NA,
  Edays,
  Ddays,
  Gdays = NA,
  DdaysO = NA,
  d,
  d_off,
  n,
  W,
  Cg = NA,
  Co = NA,
  r,
  APCcorrection = FALSE,
  DEPTH = 30,
  orig = FALSE
)
}
\arguments{
\item{SAND}{Sand fraction in the top 30 cm of soil (0 - 1)}

\item{RAIN}{Long-term Mean Annual Precipitation (mm/year).}

\item{MAT}{Long-term Mean Annual Precipitation (degrees Celsius).}

\item{FIRE}{Average number of fires per year (#/year)}

\item{LIGCELL}{Lignin and cellulose fraction of plant biomass (0 - 1)}

\item{years}{Number of years over which to run the simulation.}

\item{SOC}{Starting soil carbon stocks (t/ha).}

\item{Sk}{The steady state of biomass in the absence of grazing for a given location. This should ideally be measured directly using grazing exclosures.}

\item{S0}{Biomass condition prior to the growing season (at the end of the dry season) that is mostly comprised of carbon stores in rhizomes. Default is 0.1*SK.}

\item{Edays}{Number of days within the growing season prior to grazing episode}

\item{Ddays}{Number of days of grazing episode}

\item{Gdays}{Total number of days in the growing season. Default = 153 (October to March-ish).}

\item{DdaysO}{Number of days of offseason grazing episode. Default = 365 - Gdays.}

\item{d}{Stocking density during the growing season (head/ha)}

\item{d_off}{Stocking density during the offseason (head/ha)}

\item{n}{Number of "pastures" per total area, A.}

\item{W}{Average animal body size (kg live weight)}

\item{Cg}{Daily biomass consumption rate during the growing season (g/animal/day). Default = calc_Cg(W)}

\item{Co}{Daily biomass consumption rate during the offseason (g/animal/day). Default = Cg/2}

\item{r}{Maximum relative growth rate of grass biomass. Can be calculated as the y-intercept of a relationship between the measured relative growth rate and biomass at a given time. Default for tropical grasslands is 0.05 and for temperate grasslands is 0.035.}

\item{APCcorrection}{Default = FALSE. A correction factor for the influence of annual versus perennial plant growth strategies on belowground production. If correction is applied, then APC = 0.291. This is where forage is dominated by annuals (or shrubs often associated with annuals in drier grasslands). Otherwise, APC = 1.}

\item{DEPTH}{Default = 30. Depth of soil sampling / estimation (cm). The original SNAP model was developed based on measurements to a depth of 40 cm.}

\item{orig}{Default = FALSE. Use the original DMRESP equations from Ritchie 2020 or the updated ones from Ruan deWet.}
}
\description{
This is a wrapper function around all the SNAPGRAZE equations to calculate SOC delta over a given number of years. Management and weather inputs can be provided as either single values or as a vector/list of values representing the corresponding value for each year being modeled.
}
