% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNAPGRAZEr-package.R
\docType{package}
\name{SNAPGRAZEr-package}
\alias{SNAPGRAZEr}
\alias{SNAPGRAZEr-package}
\title{SNAPGRAZEr: A model for simulating the impacts of grazing management on SOC stocks and stock changes}
\description{
R implementation of the SNAPGRAZE model first published by Ritchie 2020 to simulate the impacts of grazing management, climate, soil, and vegetation parameters on SOC stocks and stock changes. Branched from GitHub repos originally made by 'ruan-de-wet' and 'kanedan29'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jpjazzcat/SNAPGRAZEr}
}

}
\author{
\strong{Maintainer}: Jacob Penner \email{jacob.penner@tnc.org} (\href{https://orcid.org/0000-0001-9578-4458}{ORCID})

}
\keyword{internal}
