% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_DDSOCt.R
\name{calc_DDSOCt}
\alias{calc_DDSOCt}
\title{Dung-Derived SOC in year t}
\usage{
calc_DDSOCt(LIGCELL, Ddays, Cg, n, d, Lo)
}
\arguments{
\item{LIGCELL}{Lignin and cellulose content of livestock feed for year t (\%)}

\item{Ddays}{Number of days of grazing episode}

\item{Cg}{Daily consumption rate (g/animal/day). This is the average rate during the growing season - for the dormant season, the equations assume half this consumption rate.}

\item{n}{Number of "pastures" per total area, A.}

\item{d}{Stocking density (head/ha)}

\item{Lo}{Loss of biomass during the off season. Output of calc_Lo().}
}
\description{
When it comes to soil organic carbon (SOC), there is plant-derived SOC (PDSOC) and dung-derived SOC (DDSOC).
}
