% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_deltaSOC.R
\name{calc_deltaSOC}
\alias{calc_deltaSOC}
\title{Delta SOC}
\usage{
calc_deltaSOC(PDSOCt, DDSOCt, SAND, RAIN, Gdays, SOC, orig = FALSE)
}
\arguments{
\item{PDSOCt}{Output of calc_PDSOCt()}

\item{DDSOCt}{Output of calc_DDSOCt()}

\item{SAND}{Sand fraction in the top 30 cm of soil (0 - 1)}

\item{RAIN}{MAP for year t (mm/year)}

\item{Gdays}{Total number of days in the growing season.}

\item{SOC}{Starting soil organic carbon stocks (tSOC/ha).}

\item{orig}{Default = FALSE. Use the original DMRESP equations from Ritchie 2020 or the updated ones from Ruan deWet.}
}
\description{
Calculate the change in SOC for year t (deltaSOCt) in tSOC/ha using plant-derived soil organic carbon (PDSOCt), dung-derived soil organic carbon (DDSOCt), and microbial respiration (MRESPt). A key input to MRESPt is WETDAYS, which is calculated as part of this function.
}
