% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Lo.R
\name{calc_Lo}
\alias{calc_Lo}
\title{Loss of biomass during the off season}
\usage{
calc_Lo(Cg, Gdays = 153, d)
}
\arguments{
\item{Cg}{Daily consumption rate (g/animal/day). This is the average rate during the growing season - for the dormant season, the equations assume half this consumption rate.}

\item{Gdays}{Total number of days in the growing season. Default = 153 (October to March-ish).}

\item{d}{Stocking density (head/ha)}
}
\description{
Estimating the amount of biomass lost during the dormant season (dry, non-growing season).
}
