% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_PDSOCt.R
\name{calc_PDSOCt}
\alias{calc_PDSOCt}
\title{Plant-Derived SOC in year t}
\usage{
calc_PDSOCt(BNPPest, Sf, Lo, LIGCELL, FIRE)
}
\arguments{
\item{BNPPest}{Output of calc_BNPPest()}

\item{Sf}{Biomass at the end of the growing season. Output of calc_Sf().}

\item{Lo}{Loss of biomass during the off season. Output of calc_Lo().}

\item{LIGCELL}{Lignin and cellulose content of livestock feed for year t (\%)}

\item{FIRE}{Average number of fires per year (#/year)}
}
\description{
When it comes to soil organic carbon (SOC), there is plant-derived SOC (PDSOC) and dung-derived SOC (DDSOC). PDSOCt just takes the estimated ANPP and BNPP along with LIGCELL, Fire and information about grazing management
}
