% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_SOCeq.R
\name{calc_SOCeq}
\alias{calc_SOCeq}
\title{Equilibrium SOC}
\usage{
calc_SOCeq(PDSOCt, DDSOCt, SAND, RAIN, lowSOC = FALSE)
}
\arguments{
\item{PDSOCt}{Output of calc_PDSOCt()}

\item{DDSOCt}{Output of calc_DDSOCt()}

\item{SAND}{Sand \% in top 30 cm soil}

\item{RAIN}{MAP for year t (mm/year)}

\item{lowSOC}{Default = FALSE. Different regression equation for respiration rate is applied for low and high SOC to avoid a negative respiration rate (which isn't physically possible). Threshold for what qualifies as "low SOC" is 4,600 gC/m^2 (i.e. 46 t/ha). Low SOC regression equation is applicable for higher SOC, but just with slightly lower R-squared.}
}
\description{
To calculate the change in SOC for year t (deltaSOCt), we need the combination of PDSOCt and DDSOCt, but also the maximum rate of microbial respiration for year t (MRESPt). A key input to MRESPt is WETDAYS, which is calculated as part of this function. Equilibrium SOC can be calculated by solving for deltaSOC = 0, which then avoids the need for SOC measurement.
}
