% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNAPGRAZEr-package.R
\docType{package}
\name{SNAPGRAZEr-package}
\alias{SNAPGRAZEr}
\alias{SNAPGRAZEr-package}
\title{SNAPGRAZEr: Estimate SOC using SNAPGRAZE model}
\description{
R implementation of the SNAPGRAZE model equations developed by Mark Ritchie to estimate the equilibrium soil organic carbon stocks for a given climate and rangeland management history.
}
\author{
\strong{Maintainer}: Ruan de Wet \email{ruan@meatnaturallyafrica.com} (\href{https://orcid.org/0000-0001-5221-8042}{ORCID})

}
\keyword{internal}
